serialmqtt/requirements.txt
paho-mqtt==1
pyserial==3
